refactor(tests): Tier 3 batch 3 — split dockerbuild and container tests (#192)
Split three large test files by responsibility:

- internal/dockerbuild/game_test.go (672 → 282 lines)
  → game_resolver_test.go: builder constructor + 7 resolve*/path tests
  → game_ddc_test.go: DDC arg validation, directory creation, volume format

- internal/dockerbuild/dockerfile_test.go (430 → 250 lines)
  → dockerfile_engine_test.go: full-build engine Dockerfile options, structure,
    multi-stage chain, apt/dnf packages, comment header

- internal/container/builder_test.go (313 → 162 lines)
  → builder_resolver_test.go: resolveProjectName, resolveServerTarget,
    resolveArch, resolveServerBinaryName (compute vs generate separation)

No behavior changes. All tests pass.
