+++ internal/dockerbuild/game_test.go
@@ -1,254 +1,11 @@
 package dockerbuild
 
 import (
-	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 
 	"github.com/devrecon/ludus/internal/runner"
 )
-
-func TestNewDockerGameBuilder(t *testing.T) {
-	r := runner.NewRunner(false, false)
-
-	tests := []struct {
-		name         string
-		opts         DockerGameOptions
-		wantPlatform string
-	}{
-		{
-			name:         "default platform is Linux",
-			opts:         DockerGameOptions{},
-			wantPlatform: "Linux",
-		},
-		{
-			name:         "explicit platform preserved",
-			opts:         DockerGameOptions{Platform: "Win64"},
-			wantPlatform: "Win64",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			if b.opts.Platform != tt.wantPlatform {
-				t.Errorf("Platform = %q, want %q", b.opts.Platform, tt.wantPlatform)
-			}
-		})
-	}
-}
-
-func TestResolveProjectName(t *testing.T) {
-	r := runner.NewRunner(false, false)
-
-	tests := []struct {
-		name string
-		opts DockerGameOptions
-		want string
-	}{
-		{
-			name: "defaults to Lyra",
-			opts: DockerGameOptions{},
-			want: "Lyra",
-		},
-		{
-			name: "custom project name",
-			opts: DockerGameOptions{ProjectName: "ShooterGame"},
-			want: "ShooterGame",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			got := b.resolveProjectName()
-			if got != tt.want {
-				t.Errorf("resolveProjectName() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestResolveServerTarget(t *testing.T) {
-	r := runner.NewRunner(false, false)
-
-	tests := []struct {
-		name string
-		opts DockerGameOptions
-		want string
-	}{
-		{
-			name: "defaults to LyraServer",
-			opts: DockerGameOptions{},
-			want: "LyraServer",
-		},
-		{
-			name: "custom project derives target",
-			opts: DockerGameOptions{ProjectName: "ShooterGame"},
-			want: "ShooterGameServer",
-		},
-		{
-			name: "explicit server target",
-			opts: DockerGameOptions{ServerTarget: "MyCustomServer"},
-			want: "MyCustomServer",
-		},
-		{
-			name: "explicit target overrides project name derivation",
-			opts: DockerGameOptions{ProjectName: "ShooterGame", ServerTarget: "SGServer"},
-			want: "SGServer",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			got := b.resolveServerTarget()
-			if got != tt.want {
-				t.Errorf("resolveServerTarget() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestResolveGameTarget(t *testing.T) {
-	r := runner.NewRunner(false, false)
-
-	tests := []struct {
-		name string
-		opts DockerGameOptions
-		want string
-	}{
-		{
-			name: "defaults to LyraGame",
-			opts: DockerGameOptions{},
-			want: "LyraGame",
-		},
-		{
-			name: "custom project derives target",
-			opts: DockerGameOptions{ProjectName: "ShooterGame"},
-			want: "ShooterGameGame",
-		},
-		{
-			name: "explicit game target",
-			opts: DockerGameOptions{GameTarget: "MyGameTarget"},
-			want: "MyGameTarget",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			got := b.resolveGameTarget()
-			if got != tt.want {
-				t.Errorf("resolveGameTarget() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestIsExternalProject(t *testing.T) {
-	r := runner.NewRunner(false, false)
-
-	tests := []struct {
-		name string
-		opts DockerGameOptions
-		want bool
-	}{
-		{
-			name: "no project path is not external",
-			opts: DockerGameOptions{},
-			want: false,
-		},
-		{
-			name: "with project path is external",
-			opts: DockerGameOptions{ProjectPath: "/home/user/MyGame/MyGame.uproject"},
-			want: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			got := b.isExternalProject()
-			if got != tt.want {
-				t.Errorf("isExternalProject() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestContainerProjectPath(t *testing.T) {
-	r := runner.NewRunner(false, false)
-
-	tests := []struct {
-		name string
-		opts DockerGameOptions
-		want string
-	}{
-		{
-			name: "in-engine Lyra default",
-			opts: DockerGameOptions{},
-			want: "/engine/Samples/Games/Lyra/Lyra.uproject",
-		},
-		{
-			name: "in-engine custom project",
-			opts: DockerGameOptions{ProjectName: "ShooterGame"},
-			want: "/engine/Samples/Games/ShooterGame/ShooterGame.uproject",
-		},
-		{
-			name: "external project",
-			opts: DockerGameOptions{ProjectPath: "/home/user/MyGame/MyGame.uproject", ProjectName: "MyGame"},
-			want: "/project/MyGame.uproject",
-		},
-		{
-			name: "external project defaults name to Lyra",
-			opts: DockerGameOptions{ProjectPath: "/some/path"},
-			want: "/project/Lyra.uproject",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			got := b.containerProjectPath()
-			if got != tt.want {
-				t.Errorf("containerProjectPath() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestResolveClientPlatform(t *testing.T) {
-	r := runner.NewRunner(false, false)
-
-	tests := []struct {
-		name string
-		opts DockerGameOptions
-		want string
-	}{
-		{
-			name: "defaults to Linux",
-			opts: DockerGameOptions{},
-			want: "Linux",
-		},
-		{
-			name: "explicit platform",
-			opts: DockerGameOptions{ClientPlatform: "Win64"},
-			want: "Win64",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			got := b.resolveClientPlatform()
-			if got != tt.want {
-				t.Errorf("resolveClientPlatform() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
 
 func TestScriptPreamble(t *testing.T) {
 	r := runner.NewRunner(false, false)
@@ -334,109 +91,6 @@ func TestScriptPreamble_InstallsRuntimeDeps(t *testing.T) {
 	// through a multi-hour compile only to crash at cook.
 	if !strings.Contains(got, "exit 1") {
 		t.Error("preamble must fail fast on install failure (exit 1)")
-	}
-}
-
-func TestDDCArgs(t *testing.T) {
-	r := runner.NewRunner(false, false)
-
-	tests := []struct {
-		name       string
-		ddcMode    string
-		ddcPath    string
-		wantErr    string
-		wantArgs   []string
-		wantNoArgs bool
-	}{
-		{
-			name:    "local mode with empty path errors",
-			ddcMode: "local",
-			ddcPath: "",
-			wantErr: "no path configured",
-		},
-		{
-			name:    "unsupported mode errors",
-			ddcMode: "s3",
-			ddcPath: "/some/path",
-			wantErr: "unsupported DDC mode",
-		},
-		{
-			name:       "none mode returns no args",
-			ddcMode:    "none",
-			wantNoArgs: true,
-		},
-		{
-			name:       "empty mode returns no args",
-			ddcMode:    "",
-			wantNoArgs: true,
-		},
-		{
-			name:    "local mode returns volume and env args",
-			ddcMode: "local",
-			ddcPath: filepath.Join(t.TempDir(), "ddc"),
-			wantArgs: []string{
-				"-v", // volume mount present
-				"/ddc",
-				"-e",
-				"UE-LocalDataCachePath=/ddc",
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := NewDockerGameBuilder(DockerGameOptions{DDCMode: tt.ddcMode, DDCPath: tt.ddcPath}, r)
-			checkDDCArgs(t, b, tt.wantErr, tt.wantNoArgs, tt.wantArgs)
-		})
-	}
-}
-
-func TestDDCArgs_LocalCreatesDirectory(t *testing.T) {
-	ddcDir := filepath.Join(t.TempDir(), "ddc")
-	r := runner.NewRunner(false, false)
-	b := NewDockerGameBuilder(DockerGameOptions{
-		DDCMode: "local",
-		DDCPath: ddcDir,
-	}, r)
-
-	_, err := b.ddcArgs()
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if _, err := os.Stat(ddcDir); err != nil {
-		t.Errorf("DDC directory should have been created: %v", err)
-	}
-}
-
-func TestDDCArgs_LocalVolumeFormat(t *testing.T) {
-	ddcDir := filepath.Join(t.TempDir(), "ddc")
-	r := runner.NewRunner(false, false)
-	b := NewDockerGameBuilder(DockerGameOptions{
-		DDCMode: "local",
-		DDCPath: ddcDir,
-	}, r)
-
-	args, err := b.ddcArgs()
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	// Expect exactly 4 args: -v <host>:/ddc -e UE-LocalDataCachePath=/ddc
-	if len(args) != 4 {
-		t.Fatalf("expected 4 args, got %d: %v", len(args), args)
-	}
-	if args[0] != "-v" {
-		t.Errorf("args[0] = %q, want -v", args[0])
-	}
-	if !strings.HasSuffix(args[1], ":/ddc") {
-		t.Errorf("args[1] = %q, should end with :/ddc", args[1])
-	}
-	if args[2] != "-e" {
-		t.Errorf("args[2] = %q, want -e", args[2])
-	}
-	if args[3] != "UE-LocalDataCachePath=/ddc" {
-		t.Errorf("args[3] = %q, want UE-LocalDataCachePath=/ddc", args[3])
 	}
 }
 
@@ -593,50 +247,6 @@ func TestGenerateBuildScript_ClientContainsCdEngine(t *testing.T) {
 	got := b.generateBuildScript(false)
 	if !strings.Contains(got, "cd /engine") {
 		t.Errorf("client build script should contain 'cd /engine'\ngot:\n%s", got)
-	}
-}
-
-func TestDDCNotInBuildScript(t *testing.T) {
-	tests := []struct {
-		name string
-		opts DockerGameOptions
-	}{
-		{
-			name: "local mode does not embed DDC args in script",
-			opts: DockerGameOptions{DDCMode: "local", DDCPath: "/tmp/ddc"},
-		},
-		{
-			name: "none mode has no DDC args in script",
-			opts: DockerGameOptions{DDCMode: "none"},
-		},
-		{
-			name: "empty mode has no DDC args in script",
-			opts: DockerGameOptions{},
-		},
-	}
-
-	notExpected := []string{"-ini:Engine:[DerivedDataBackendGraph]", "UE-LocalDataCachePath"}
-
-	r := runner.NewRunner(false, false)
-	for _, tt := range tests {
-		t.Run(tt.name+" (server)", func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			got := b.generateBuildScript(true)
-			for _, notWant := range notExpected {
-				if strings.Contains(got, notWant) {
-					t.Errorf("server script should NOT contain %q\ngot:\n%s", notWant, got)
-				}
-			}
-		})
-		t.Run(tt.name+" (client)", func(t *testing.T) {
-			b := NewDockerGameBuilder(tt.opts, r)
-			got := b.generateBuildScript(false)
-			for _, notWant := range notExpected {
-				if strings.Contains(got, notWant) {
-					t.Errorf("client script should NOT contain %q\ngot:\n%s", notWant, got)
-				}
-			}
-		})
 	}
 }
